feat(agc): execute indirect multi-draw packets

+++ src/SharpEmu.Libs/Agc/AgcIndirectDrawArguments.cs
@@ -16,7 +16,7 @@ internal readonly record struct AgcIndirectDrawArguments(
     public static bool TryRead(
         CpuContext context,
         ulong argumentBufferAddress,
-        uint byteOffset,
+        ulong byteOffset,
         bool indexed,
         out AgcIndirectDrawArguments arguments)
     {
